Delete the pickup  and request associated by this pickup with contraints

+++ src/main/java/tn/workbot/coco_marketplace/services/PickupService.java
@@ -2,12 +2,14 @@ package tn.workbot.coco_marketplace.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.webjars.NotFoundException;
 import tn.workbot.coco_marketplace.entities.*;
 import tn.workbot.coco_marketplace.entities.enmus.StatusPickupBuyer;
 import tn.workbot.coco_marketplace.entities.enmus.StatusPickupSeller;
 import tn.workbot.coco_marketplace.repositories.*;
 import tn.workbot.coco_marketplace.services.interfaces.PickupIService;
 
+import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -35,31 +37,35 @@ public class PickupService implements PickupIService {
         int randomNumber = random.nextInt(9000) + 1000;  // generates a random number between 1000 and 9999
         String prefix = "216";
         String code = prefix + randomNumber;
-        List<Pickup>pickups= (List<Pickup>) pr.findAll();
-        for (Pickup p:pickups) {
-                      if(p.getCodePickup()!=code){
-                          pickup.setCodePickup(code);
-                      }
-                      else {
-                          int randomNumber1 = random.nextInt(100) + 100;
-                          String code1 = prefix + randomNumber + randomNumber1;
-                          pickup.setCodePickup(code1);
-                      }
+        List<Pickup> pickups = (List<Pickup>) pr.findAll();
+        for (Pickup p : pickups) {
+            if (p.getCodePickup() != code) {
+                pickup.setCodePickup(code);
+            } else {
+                int randomNumber1 = random.nextInt(100) + 100;
+                String code1 = prefix + randomNumber + randomNumber1;
+                pickup.setCodePickup(code1);
+            }
         }
         pickup.setCodePickup(code);
         pickup.setDateCreationPickup(LocalDateTime.now());
         return pr.save(pickup);
     }
-
+    @Transactional
     @Override
     public void removePickup(Long id) {
-         pr.deleteById(id);
+        Pickup pickup = pr.findById(id).orElseThrow(() -> new NotFoundException("Pickup not found"));
+        String PICKED = "PICKED";
+        if (!pickup.getStatusPickupSeller().equals(StatusPickupSeller.PICKED)) {
+            throw new IllegalStateException("Cannot remove pickup with status other than PICKED");
+        }
+        pr.deleteById(id);
     }
 
     @Override
     public Pickup RetrievePickup(Long id) {
-        if(pr.findById(id).isPresent())
-        return pr.findById(id).get();
+        if (pr.findById(id).isPresent())
+            return pr.findById(id).get();
         return new Pickup();
     }
 
@@ -80,8 +86,8 @@ public class PickupService implements PickupIService {
 
     @Override
     public Pickup AssignPickupByOder(Pickup pickup, Long id) {
-        Pickup p=pr.save(pickup);
-        Order order=or.findById(id).get();
+        Pickup p = pr.save(pickup);
+        Order order = or.findById(id).get();
         p.setOrder(order);
         return pr.save(pickup);
     }
@@ -90,54 +96,51 @@ public class PickupService implements PickupIService {
     public List<Pickup> RetrievePickupsByGovernoratBetweenStoreAndDeliveryMenFreelancer() {
         //HADHI session variable
         User u = ur.findById(1L).get();
-        List<Store> store= (List<Store>) sr.findAll();
+        List<Store> store = (List<Store>) sr.findAll();
         List<Store> storesInSameGovernorate = sr.findByGovernorate(u.getGovernorate());
         List<Pickup> pickups = new ArrayList<>();
         //ELI 3ANDHOUM BIKE MATODHA7RELHOUM KEN EL PICKUPS ELI BECH TTLVRA FEN HOUMA W STORE YABDA FARD GOVERNORATE
-       if((u.getGear().equals("BIKE")) || (u.getGear().equals("MOTO")))
-       {
-           for (Store s : store) {
-               if (u.getGovernorate().equals(s.getGovernorate())) {
-                   return pr.findByGovernorate(s.getGovernorate());
-               }
-           }
-       }
-       else if(u.getGear().equals("CAR"))
-       {
-           for (Store storee : storesInSameGovernorate) {
-               pickups.addAll(storee.getPickups());
-           }
+        if ((u.getGear().equals("BIKE")) || (u.getGear().equals("MOTO"))) {
+            for (Store s : store) {
+                if (u.getGovernorate().equals(s.getGovernorate())) {
+                    return pr.findByGovernorate(s.getGovernorate());
+                }
+            }
+        } else if (u.getGear().equals("CAR")) {
+            for (Store storee : storesInSameGovernorate) {
+                pickups.addAll(storee.getPickups());
+            }
 
-           return pickups;
-       }
-        return  Collections.emptyList();
+            return pickups;
+        }
+        return Collections.emptyList();
     }
 
 
     @Override
     public List<Pickup> RetrievePickupsbetweenAgencyBranchAndStoreInTheSomeGovernorat() {
         User u = ur.findById(1L).get();
-        List<AgencyBranch> agencyBranch=abr.findAll();
-        List<User> users= (List<User>) ur.findAll();
-        List<Store>stores=sr.findAll();
+        List<AgencyBranch> agencyBranch = abr.findAll();
+        List<User> users = (List<User>) ur.findAll();
+        List<Store> stores = sr.findAll();
         List<AgencyBranch> agencyBranches = new ArrayList<>();
         ArrayList<Pickup> store = new ArrayList<>();
         agencyBranches.addAll(u.getAgencyBranches());
-        for (AgencyBranch ab:agencyBranches) {
-            for(Store s:stores){
-                if(s.getGovernorate().equals(ab.getGovernorate())){
+        for (AgencyBranch ab : agencyBranches) {
+            for (Store s : stores) {
+                if (s.getGovernorate().equals(ab.getGovernorate())) {
                     store.addAll(s.getPickups());
                 }
             }
         }
-        return store ;
+        return store;
     }
 
     @Override
     public Pickup AssignPickupByStore(Pickup pickup) {
         //Variable Of Session Manager
-        Store store=sr.findById(1L).get();
-        Pickup pickup1=pr.save(pickup);
+        Store store = sr.findById(1L).get();
+        Pickup pickup1 = pr.save(pickup);
         pickup1.setStore(store);
         Random random = new Random(); //java.util.Random
         pickup.setStatusPickupSeller(StatusPickupSeller.valueOf("PICKED"));
@@ -145,12 +148,11 @@ public class PickupService implements PickupIService {
         int randomNumber = random.nextInt(9000) + 1000;  // generates a random number between 1000 and 9999
         String prefix = "216";
         String code = prefix + randomNumber;
-        List<Pickup>pickups= (List<Pickup>) pr.findAll();
-        for (Pickup p:pickups) {
-            if(p.getCodePickup()!=code){
+        List<Pickup> pickups = (List<Pickup>) pr.findAll();
+        for (Pickup p : pickups) {
+            if (p.getCodePickup() != code) {
                 pickup.setCodePickup(code);
-            }
-            else {
+            } else {
                 int randomNumber1 = random.nextInt(100) + 100;
                 String code1 = prefix + randomNumber + randomNumber1;
                 pickup.setCodePickup(code1);
@@ -162,24 +164,21 @@ public class PickupService implements PickupIService {
     }
 
     @Override
-    public Pickup ModifyStatusOfPickupByDelivery(String Status,Long idPickup) {
-        Pickup pickup=pr.findById(idPickup).get();
-        if(Status.equals("SHIPPED")){
+    public Pickup ModifyStatusOfPickupByDelivery(String Status, Long idPickup) {
+        Pickup pickup = pr.findById(idPickup).get();
+        if (Status.equals("SHIPPED")) {
             pickup.setStatusPickupBuyer(StatusPickupBuyer.valueOf("SHIPPED"));
             pickup.setStatusPickupSeller(StatusPickupSeller.valueOf("ONTHEWAY"));
             pr.save(pickup);
-        }
-        else if (Status.equals("DELIVERED")){
+        } else if (Status.equals("DELIVERED")) {
             pickup.setStatusPickupBuyer(StatusPickupBuyer.valueOf("DELIVERED"));
             pickup.setStatusPickupSeller(StatusPickupSeller.valueOf("DELIVERED"));
             pr.save(pickup);
-        }
-        else if (Status.equals("RETURN")){
+        } else if (Status.equals("RETURN")) {
             pickup.setStatusPickupBuyer(StatusPickupBuyer.valueOf("RETURN"));
             pickup.setStatusPickupSeller(StatusPickupSeller.valueOf("RETURN"));
             pr.save(pickup);
-        }
-        else {
+        } else {
             pickup.setStatusPickupBuyer(StatusPickupBuyer.valueOf("REFUNDED"));
             pickup.setStatusPickupSeller(StatusPickupSeller.valueOf("REFUNDED"));
             pr.save(pickup);
